requirements.txt
pyqrcode
Pillow
opencv-python
mysql-connector-python
